requirements.txt
tqdm==4.43.0
tensorflow_gpu==2.0.0
scipy==1.5.4
numpy==1.19.5
pandas==1.1.5
